rockchip: test: fix size warning in arm64
Fix warning in emmc and nand test case.

+++ test/rockchip/test-emmc.c
@@ -69,7 +69,7 @@ int board_emmc_test(int argc, char * const argv[])
 	/* 2. Prepare and start cli command */
 
 	snprintf(cmd_mmc, sizeof(cmd_mmc), "mmc write 0x%x 0x%x 0x%x",
-		 (u32)write_buffer, lba, blocks);
+		 (u32)(ulong)write_buffer, lba, blocks);
 	ts = get_timer(0);
 	err = cli_simple_run_command(cmd_mmc, 0);
 	ts = get_timer(0) - ts;
@@ -80,7 +80,7 @@ int board_emmc_test(int argc, char * const argv[])
 		blocks / 2048, ts, (blocks >> 1) / ts);
 
 	snprintf(cmd_mmc, sizeof(cmd_mmc), "mmc read 0x%x 0x%x 0x%x",
-		 (u32)read_buffer, lba, blocks);
+		 (u32)(ulong)read_buffer, lba, blocks);
 	ts = get_timer(0);
 	err = cli_simple_run_command(cmd_mmc, 0);
 	ts = get_timer(0) - ts;
